Adds externalFundAccount() for Goldsand OPERATOR only (#2)
* Adds fundAccount() for Goldsand OPERATOR only

fundAccount() allows Goldsand OPERATOR to deposit funds on behalf of another account. This helps Goldsand for many use-cases, but the primary use-case is to transfer existing staking funds into the smart contract

* Fixes tests and remove minimum for externally funding an account

Renamed fundAccount() => externalFund() to align with existing naming conventions.

Removed minimum for funding an account. This is because the OPERATOR is not subject to the minimum deposit amount for an external fund